respect DESTDIR and dont use -s with install
$(DESTDIR) is the standard for installing into other trees, not $(BASEDIR) ...
so I've converted the Makefile to use that.  I've also left in $(BASEDIR) as a
default to support old installs; not sure if you'd just cut it.

Stripping should be left to the person to handle, not automatically done by
the install step.  Also, `install -s` always calls `strip` which is
wrong/undesired in cross-compiling scenarios.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Neil Brown <neilb@suse.de>